merge/generics: refactor examples&add TODO
