+++ examples/generics/__init__.py
@@ -1,0 +1,2 @@
+
+# TODO: add telegram as alter option for generics testing with Attachments and etc
