+++ examples/generics/_example_utils.py
