+++ tests/generics/test_examples.py
@@ -5,7 +5,7 @@ import pytest
 import tests.utils as utils
 
 dot_path_to_addon = utils.get_dot_path_from_tests_to_current_dir(__file__)
-example_utils = importlib.import_module(f"examples.{dot_path_to_addon}.example_utils")
+example_utils = importlib.import_module(f"examples.{dot_path_to_addon}._example_utils")
 
 @pytest.mark.parametrize(
     "example_module_name", [
